adapted Login.ts to flux
